pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fiscalpilot"
version = "0.1.0"
description = "The Open-Source AI Financial Copilot — Find waste, fraud & abuse. Maximize margins. Cut costs. Your AI-powered CFO."
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
authors = [
    { name = "FiscalPilot Contributors" },
]
keywords = [
    "ai",
    "finance",
    "agent",
    "fraud-detection",
    "cost-optimization",
    "waste-detection",
    "cfo",
    "financial-analysis",
    "llm",
    "open-source",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

dependencies = [
    "litellm>=1.40.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "typer>=0.12",
    "pandas>=2.0",
    "openpyxl>=3.1",
    "sqlalchemy>=2.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "cryptography>=42.0",
]

[project.optional-dependencies]
quickbooks = ["python-quickbooks>=0.9"]
xero = ["xero-python>=4.0"]
plaid = ["plaid-python>=20.0"]
ocr = ["pytesseract>=0.3", "pdf2image>=1.16", "Pillow>=10.0"]
all = [
    "python-quickbooks>=0.9",
    "xero-python>=4.0",
    "plaid-python>=20.0",
    "pytesseract>=0.3",
    "pdf2image>=1.16",
    "Pillow>=10.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
]

[project.urls]
Homepage = "https://github.com/fiscalpilot/fiscalpilot"
Documentation = "https://fiscalpilot.dev"
Repository = "https://github.com/fiscalpilot/fiscalpilot"
Issues = "https://github.com/fiscalpilot/fiscalpilot/issues"
Changelog = "https://github.com/fiscalpilot/fiscalpilot/blob/main/CHANGELOG.md"

[project.scripts]
fiscalpilot = "fiscalpilot.cli:app"
fp = "fiscalpilot.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/fiscalpilot"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM", "TCH"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
